# openapi/wsapi-api.yml
openapi: 3.0.3
info:
  title: WSAPI - WhatsApp API
  description: RESTful HTTP API for WhatsApp via whatsmeow. Supports multi-instance connections, webhook event delivery, and HMAC-SHA256 signing.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - InstanceAuth: []
paths:
...
  /admin/instances/{id}:
    get:
      tags: [Instances]
      summary: Get instance details
      operationId: getInstance
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdPath"
      responses:
        "200":
          description: Instance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Instances]
      summary: Delete an instance
      operationId: deleteInstance
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/InstanceIdPath"
      responses:
        "204":
          description: Instance deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /users/me/profile:
    get:
      tags: [Users]
      summary: Get own profile info
      operationId: getMyProfile
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: Account profile info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMeInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    put:
      tags: [Users]
      summary: Update own profile
      description: Update name, status, and/or picture. All fields are optional.
      operationId: updateMyProfile
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "204":
          description: Profile updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /users/me/presence:
    put:
      tags: [Users]
      summary: Set presence state
      description: Set the account's presence to available or unavailable. This is a transient action (not persisted in the profile).
      operationId: setPresence
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMyPresenceRequest"
      responses:
        "204":
          description: Presence updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /users/me/privacy:
    get:
      tags: [Users]
      summary: Get privacy settings
      description: Returns the current privacy settings for the account.
      operationId: getPrivacySettings
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: Current privacy settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacySettingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    put:
      tags: [Users]
      summary: Update a privacy setting
      description: Update a single privacy setting. Returns the full updated privacy settings.
      operationId: setPrivacySetting
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPrivacyRequest"
            examples:
              lastSeen:
                summary: Set last seen to contacts only
                value:
                  setting: "last"
                  value: "contacts"
              readReceipts:
                summary: Disable read receipts
                value:
                  setting: "readreceipts"
                  value: "none"
      responses:
        "200":
          description: Updated privacy settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacySettingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /users/check:
    post:
      tags: [Users]
      summary: Bulk check phone numbers on WhatsApp
      description: Check multiple phone numbers at once to see if they are registered on WhatsApp.
      operationId: bulkCheckUsers
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCheckRequest"
            example:
              phones: ["1234567890", "9876543210"]
      responses:
        "200":
          description: Bulk check results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BulkCheckResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /users/{phone}/check:
    get:
      tags: [Users]
      summary: Check if a phone number is on WhatsApp
      operationId: checkUser
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/PhonePath"
      responses:
        "200":
          description: WhatsApp presence check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  isInWhatsApp:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /users/{phone}/profile:
    get:
      tags: [Users]
      summary: Get user profile info
      operationId: getUserProfile
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/PhonePath"
      responses:
        "200":
          description: User profile info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ─── Media ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Admin API key for instance management endpoints (`/admin/instances/*`). Set via `auth.adminApiKey` config.

    InstanceAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Instance API key. Resolved from the instance's own `apiKey` or the global `auth.defaultApiKey`.
  parameters:
    XInstanceId:
      name: X-Instance-Id
      in: header
      required: true
      description: The ID of the WhatsApp instance to use
      schema:
        type: string
        default: "{{instanceId}}"
    InstanceIdPath:
      name: id
      in: path
      required: true
      description: Instance ID
      schema:
        type: string
...
    PhonePath:
      name: phone
      in: path
      required: true
      description: Phone number (e.g. 1234567890)
      schema:
        type: string
        pattern: '^\+?\d{7,15}$'
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 400
            detail: "invalid request body"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 401
            detail: "unauthorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 404
            detail: "not found"
    ServiceUnavailable:
      description: Instance not available (not connected or service is nil)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 503
            detail: "instance not available"
    DeviceNotPaired:
      description: Device is not paired. Complete the pairing process first.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 403
            detail: "device not paired"
  schemas:
    Error:
      type: object
      description: Standard error response returned by all endpoints on failure.
      required: [status, detail]
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable error message
          example: "invalid request body"

    # ─── Instance Schemas ──────────────────────────────────────────
...
    InstanceConfig:
      type: object
      properties:
        apiKey:
          type: string
        webhookUrl:
          type: string
          format: uri
        signingSecret:
          type: string
        eventFilters:
          type: array
          items:
            type: string
            enum:
              - chat_setting
              - chat_presence
              - chat_push_name
              - chat_status
              - chat_picture
              - message
              - message_read
              - message_delete
              - message_star
              - message_history_sync
              - contact
              - group
              - call_offer
              - call_terminate
              - call_accept
              - user_presence
          description: >-
            List of event types to deliver (empty = all events).
            System events (`logged_in`, `logged_out`, `login_error`,
            `initial_sync_finished`) are always delivered regardless of this
            filter and are silently stripped if included in the list.
        historySync:
          type: boolean
          nullable: true
...
    InstanceListItem:
      type: object
      properties:
        id:
          type: string
        config:
          $ref: "#/components/schemas/InstanceConfig"
        isConnected:
          type: boolean
        isLoggedIn:
          type: boolean
        deviceId:
          type: string

    # ─── Session Schemas ───────────────────────────────────────────
...
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          description: Display name
        status:
          type: string
          description: Status text
        picture:
          type: string
          description: Base64-encoded JPEG profile picture
    SetMyPresenceRequest:
      type: object
      required: [presence]
      properties:
        presence:
          type: string
          enum: [available, unavailable]
          description: Presence state

    # ─── Call Schemas ──────────────────────────────────────────────

    # ─── Identity Schemas ────────────────────────────────────────
    Identity:
      type: object
      description: >
        Represents a WhatsApp user with all known identifiers. The `id` field
        is always present: it prefers a phone-based JID, falling back to LID.
      properties:
        id:
          type: string
          description: Primary identifier (phone-based JID preferred, LID fallback)
          example: "1234567890@s.whatsapp.net"
        lid:
          type: string
          description: LID-based JID (omitted if unknown)
          example: "222424242@lid"
        phone:
          type: string
          description: Phone number without prefix (omitted if unknown)
          example: "1234567890"
        device:
          type: integer
          description: Device number (omitted if 0)
          example: 2
...
    UserInfo:
      allOf:
        - $ref: "#/components/schemas/Identity"
        - type: object
          properties:
            isInWhatsApp:
              type: boolean
            status:
              type: string
            pictureId:
              type: string
            pictureUrl:
              type: string
              description: Profile picture URL (omitted if unavailable)
            isVerified:
              type: boolean
    UserMeInfo:
      allOf:
        - $ref: "#/components/schemas/Identity"
        - type: object
          properties:
            deviceId:
              type: integer
            pushName:
              type: string
            businessName:
              type: string
            status:
              type: string
            pictureId:
              type: string
            isVerified:
              type: boolean
...
    BulkCheckRequest:
      type: object
      required: [phones]
      properties:
        phones:
          type: array
          items:
            type: string
          minItems: 1
          description: List of phone numbers to check
    BulkCheckResult:
      type: object
      properties:
        query:
          type: string
          description: The phone number that was queried
        isInWhatsApp:
          type: boolean
          description: Whether the phone number is registered on WhatsApp
        jid:
          type: string
          description: JID of the user (only present if registered)

    # ─── Privacy Schemas ──────────────────────────────────────────
    SetPrivacyRequest:
      type: object
      required: [setting, value]
      properties:
        setting:
          type: string
          enum: [groupadd, last, status, profile, readreceipts, online, calladd]
          description: The privacy setting to update
        value:
          type: string
          enum: [all, contacts, contact_blacklist, match_last_seen, known, none]
          description: The privacy level to set
    PrivacySettingsResponse:
      type: object
      properties:
        groupAdd:
          type: string
          description: Who can add you to groups
        lastSeen:
          type: string
          description: Who can see your last seen
        status:
          type: string
          description: Who can see your status
        profile:
          type: string
          description: Who can see your profile picture
        readReceipts:
          type: string
          description: Whether read receipts are enabled
        online:
          type: string
          description: Who can see when you are online
        callAdd:
          type: string
          description: Who can call you

    # ─── Newsletter Schemas ─────────────────────────────────────────
...